I should commit more often :)
Changed many things:
-encode and decode object functions to save/load data
-can set object fill/stroke properly
-object selections are not terribly ugly
-canvas background is new color cause white looked bad

*text selection still broken, but kinda works
*bugs regarding inconsistent/erratic panning, especially right after zooming

Still want to add an infinite background grid

+++ app/src/main/java/cornjob/flowsketch/ColorPicker.java
@@ -81,8 +81,10 @@ public class ColorPicker extends View {
 
     private Matrix gradientRotationMatrix;
 
-    /** Currently selected color */
-    private float[] colorHSV = new float[] { 0f, 0f, 1f };
+    /**
+     * Currently selected color
+     */
+    private float[] colorHSV = new float[]{0f, 0f, 1f};
 
     public ColorPicker(Context context, AttributeSet attrs, int defStyle) {
         super(context, attrs, defStyle);
@@ -160,9 +162,9 @@ public class ColorPicker extends View {
 
         // drawing value slider
 
-        float[] hsv = new float[] { colorHSV[0], colorHSV[1], 1f };
+        float[] hsv = new float[]{colorHSV[0], colorHSV[1], 1f};
 
-        SweepGradient sweepGradient = new SweepGradient(centerX, centerY, new int[] { Color.BLACK, Color.HSVToColor(hsv), Color.WHITE }, null);
+        SweepGradient sweepGradient = new SweepGradient(centerX, centerY, new int[]{Color.BLACK, Color.HSVToColor(hsv), Color.WHITE}, null);
         sweepGradient.setLocalMatrix(gradientRotationMatrix);
         valueSliderPaint.setShader(sweepGradient);
 
@@ -183,7 +185,7 @@ public class ColorPicker extends View {
 
         // drawing value pointer
 
-        valuePointerPaint.setColor(Color.HSVToColor(new float[] { 0f, 0f, 1f - colorHSV[2] }));
+        valuePointerPaint.setColor(Color.HSVToColor(new float[]{0f, 0f, 1f - colorHSV[2]}));
 
         double valueAngle = (colorHSV[2] - 0.5f) * Math.PI;
         float valueAngleX = (float) Math.cos(valueAngle);
@@ -239,10 +241,10 @@ public class ColorPicker extends View {
         int centerX = width / 2;
         int centerY = height / 2;
 
-        innerPadding = (int) (paramInnerPadding * width / 100);
-        outerPadding = (int) (paramOuterPadding * width / 100);
-        arrowPointerSize = (int) (paramArrowPointerSize * width / 100);
-        valueSliderWidth = (int) (paramValueSliderWidth * width / 100);
+        innerPadding = paramInnerPadding * width / 100;
+        outerPadding = paramOuterPadding * width / 100;
+        arrowPointerSize = paramArrowPointerSize * width / 100;
+        valueSliderWidth = paramValueSliderWidth * width / 100;
 
         outerWheelRadius = width / 2 - outerPadding - arrowPointerSize;
         innerWheelRadius = outerWheelRadius - valueSliderWidth;
@@ -271,7 +273,7 @@ public class ColorPicker extends View {
         int colorCount = 12;
         int colorAngleStep = 360 / 12;
         int colors[] = new int[colorCount + 1];
-        float hsv[] = new float[] { 0f, 1f, 1f };
+        float hsv[] = new float[]{0f, 1f, 1f};
         for (int i = 0; i < colors.length; i++) {
             hsv[0] = (i * colorAngleStep + 180) % 360;
             colors[i] = Color.HSVToColor(hsv);
@@ -295,30 +297,30 @@ public class ColorPicker extends View {
     public boolean onTouchEvent(MotionEvent event) {
         int action = event.getAction();
         switch (action) {
-        case MotionEvent.ACTION_DOWN:
-        case MotionEvent.ACTION_MOVE:
+            case MotionEvent.ACTION_DOWN:
+            case MotionEvent.ACTION_MOVE:
 
-            int x = (int) event.getX();
-            int y = (int) event.getY();
-            int cx = x - getWidth() / 2;
-            int cy = y - getHeight() / 2;
-            double d = Math.sqrt(cx * cx + cy * cy);
+                int x = (int) event.getX();
+                int y = (int) event.getY();
+                int cx = x - getWidth() / 2;
+                int cy = y - getHeight() / 2;
+                double d = Math.sqrt(cx * cx + cy * cy);
 
-            if (d <= colorWheelRadius) {
+                if (d <= colorWheelRadius) {
 
-                colorHSV[0] = (float) (Math.toDegrees(Math.atan2(cy, cx)) + 180f);
-                colorHSV[1] = Math.max(0f, Math.min(1f, (float) (d / colorWheelRadius)));
+                    colorHSV[0] = (float) (Math.toDegrees(Math.atan2(cy, cx)) + 180f);
+                    colorHSV[1] = Math.max(0f, Math.min(1f, (float) (d / colorWheelRadius)));
 
-                invalidate();
+                    invalidate();
 
-            } else if (x >= getWidth() / 2 && d >= innerWheelRadius) {
+                } else if (x >= getWidth() / 2 && d >= innerWheelRadius) {
 
-                colorHSV[2] = (float) Math.max(0, Math.min(1, Math.atan2(cy, cx) / Math.PI + 0.5f));
+                    colorHSV[2] = (float) Math.max(0, Math.min(1, Math.atan2(cy, cx) / Math.PI + 0.5f));
 
-                invalidate();
-            }
+                    invalidate();
+                }
 
-            return true;
+                return true;
         }
         return super.onTouchEvent(event);
     }
